Make use of generic names
