+++ jest.config.js
@@ -13,7 +13,7 @@ module.exports = {
     'src/components/login/*.test.js',
     'src/components/newsFeed/index.test.js', // This component doesn't meet the setted tresholds for mocha but in jest
     'src/components/screens/register/register.test.js',
-    'src/components/screens/delegates/voting/voteUrlProcessor/index.test.js',
+    'src/components/screens/delegates/votingSummary/voteUrlProcessor/index.test.js',
     'src/store/reducers/liskService.test.js',
     'src/components/screens/register/register.test.js',
     'src/components/shared/header/signInHeader/signInHeader.test.js',
@@ -54,7 +54,7 @@ module.exports = {
     'src/components/shared/transactionDashboard/index.js',
     'src/components/screens/wallet/transactions/transactions.js',
     'src/components/screens/explorer/transactions/transactions.js',
-    'src/components/screens/delegates/voting/voteUrlProcessor/index.js',
+    'src/components/screens/delegates/votingSummary/voteUrlProcessor/index.js',
     'src/constants/',
     'src/i18n-scanner.js',
     'src/main.js',
@@ -71,7 +71,7 @@ module.exports = {
     'src/utils/theme.js',
     'src/components/screens/registerDelegate/registerDelegate.js',
     'src/components/wallet/transactions/transactionsOverview.js',
-    'src/components/screens/delegates/voting/voting.js',
+    'src/components/screens/delegates/votingSummary/voting.js',
     'src/components/screens/explorer/transactions/index.js',
     'src/components/screens/explorer/accounts/accounts.js',
     'src/components/toolbox/dropdown/toolBoxDropdown.js',
@@ -86,7 +86,7 @@ module.exports = {
     'src/components/shared/header/signInHeader/signInHeader.js',
     'src/components/screens/register/register.js',
     'src/components/toolbox/demoRenderer.js',
-    'src/components/screens/delegates/votingHeader.js',
+    'src/components/screens/delegates/header.js',
     'src/components/screens/monitor/delegate/delegate.js',
     'src/components/screens/monitor/delegates/delegateRow.js',
     'src/components/screens/monitor/delegates/delegates.js',
