+++ src/components/screens/delegates/table/delegateRow.js
