+++ src/components/screens/delegates/table/delegateRow.test.js
