+++ src/components/screens/delegates/table/delegatesTable.css
