+++ src/components/screens/delegates/table/delegatesTable.test.js
