+++ src/components/screens/delegates/table/voteCheckbox.js
