+++ src/components/screens/delegates/table/withDelegatesData.js
