+++ src/components/screens/delegates/votingSummary/index.js
