+++ src/components/screens/delegates/votingSummary/voteList.css
