+++ src/components/screens/delegates/votingSummary/voteList.js
