+++ src/components/screens/delegates/votingSummary/voteUrlProcessor/index.js
