+++ src/components/screens/delegates/votingSummary/voteUrlProcessor/voteUrlProcessor.js
