+++ src/components/screens/delegates/votingSummary/voteUrlProcessor/voteUrlProcessor.test.js
