+++ src/components/screens/delegates/votingSummary/voting.css
