+++ src/components/screens/delegates/votingSummary/voting.js
