+++ src/components/screens/delegates/votingSummary/voting.test.js
