+++ src/components/screens/delegates/votingSummary/votingSummary.js
